Add some new code to handle D90 commands
I've pulled the card out of my Nikon D90 a few times while
it is accessing the card.  I can then put it in the PC and
see what the D90 was trying to do.

I've noticed two new commands: one to disable the WiFi on
the card, and the other to check to see if the WiFi is
currently disabled or not.

I haven't hooked these up to the user interface yet.

I also think the 'o' 13 command tells which ESSID the card
is currently associated with.  'o' 12 might be something to
do with pending uploads.