add content searching for key type detection
This used to just simply use the number of bytes in the argument
as the method to detect which type of wifi encryption the network
is using.  That sucked if you happened to use a 10-character
(WEP_KEY_BYTES*2) password.  This at least makes sure to not treat
it as hex unless it's all hex characters.